requirements.txt
# Core Django
Django>=5.0,<5.1
django-cors-headers>=4.3.0
whitenoise>=6.6.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# AI Integration
google-genai>=0.3.0
pyaudio>=0.2.14
openai>=1.3.0

# API
djangorestframework>=3.14.0
django-filter>=23.3

# Background Tasks
celery>=5.3.4
redis>=5.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Environment Management
python-dotenv>=1.0.0

# Security & Authentication
cryptography>=41.0.5

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

